Introduce dashboard modules for Git operations and path resolution
This commit adds new modules for handling Git operations and path resolution within the dashboard API. The `dashboard_git` module includes functions for managing commit logs, user filtering, and parsing commit statistics. The `dashboard_paths` module provides utilities for resolving bundle files and determining content types based on file extensions. Additionally, the `dashboard_runtime` module is updated to integrate these new functionalities, enhancing the overall structure and maintainability of the dashboard API. Tests are also updated to reflect these changes, ensuring robust functionality.
